refactor: create `intervals` module
Soon we'll be supporting stop intervals in addition to missing intervals,
but much of the code will be shared.
